# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portfolio Next API Backend
  version: 1.0.0
  description: Rebuilt backend APIs (MongoDB Atlas + Next.js route handlers)
servers:
  - url: /
security: []
...
paths:
...
  /api/admin/content/home:
    get:
      tags: [Admin Content]
      summary: Admin get home content
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Home content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HomeResponse' }
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      tags: [Admin Content]
      summary: Admin patch home content
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HomeUpdateInput'
      responses:
        '200':
          description: Updated home content
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HomeResponse' }
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /api/admin/content/projects:
    get:
      tags: [Admin Content]
      summary: Admin list projects
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Projects list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectsResponse' }
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      tags: [Admin Content]
      summary: Admin create project
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectCreateInput' }
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
  /api/admin/content/projects/{id}:
    get:
      tags: [Admin Content]
      summary: Admin get project
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Project document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      tags: [Admin Content]
      summary: Admin patch project
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProjectPatchInput' }
      responses:
        '200':
          description: Updated project
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProjectResponse' }
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags: [Admin Content]
      summary: Admin delete project
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteSuccess' }
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
  /api/admin/content/posts:
    get:
      tags: [Admin Content]
      summary: Admin list posts
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Posts list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostsResponse' }
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
    post:
      tags: [Admin Content]
      summary: Admin create post
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostCreateInput' }
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostResponse' }
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '500':
          $ref: '#/components/responses/Error500'
  /api/admin/content/posts/{id}:
    get:
      tags: [Admin Content]
      summary: Admin get post
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Post document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostResponse' }
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    put:
      tags: [Admin Content]
      summary: Admin patch post
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PostPatchInput' }
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PostResponse' }
        '400':
          $ref: '#/components/responses/Error400'
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
    delete:
      tags: [Admin Content]
      summary: Admin delete post
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeleteSuccess' }
        '401':
          $ref: '#/components/responses/Error401'
        '404':
          $ref: '#/components/responses/Error404'
        '500':
          $ref: '#/components/responses/Error500'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Error404:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Error500:
      description: Internal error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success: { type: boolean, enum: [false] }
        error: { type: string }
    DeleteSuccess:
      type: object
      required: [success]
      properties:
        success: { type: boolean, enum: [true] }
...
    ImageRef:
      type: object
      required: [permalink]
      properties:
        permalink: { type: string }
    SkillEntry:
      type: object
      required: [id, skill_name]
      properties:
        id: { type: string }
        skill_name: { type: string }
    HomeContent:
      type: object
      required:
        - key
        - name
        - role
        - description
        - social_links
        - buttons
        - featured_projects_section_1title
        - featured_projects_section_1project_description
        - experience_section_1title
        - experience_section_1description
        - experience_section_1replicator_field
        - technology_section_1section_title
        - technology_section_1description
        - technology_section_1skills
        - post_sectiontitle
        - post_sectiondescription
        - seo_settings
        - createdAt
        - updatedAt
      properties:
        key: { type: string, enum: [default] }
        name: { type: string }
        role: { type: string }
        description: { type: string }
        social_links:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              social_icon: { $ref: '#/components/schemas/ImageRef' }
              social_link: { type: string }
        buttons:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              buttton_text: { type: string }
              button_link: { type: string }
              fill_background: { type: boolean }
        featured_projects_section_1title: { type: string }
        featured_projects_section_1project_description: { type: string }
        experience_section_1title: { type: string }
        experience_section_1description: { type: string }
        experience_section_1replicator_field:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              company_name: { type: string }
              job_title: { type: string }
              job_description: { type: string }
              from: { type: string }
              to: { type: string }
              present: { type: boolean }
        technology_section_1section_title: { type: string }
        technology_section_1description: { type: string }
        technology_section_1skills:
          type: array
          items: { $ref: '#/components/schemas/SkillEntry' }
        post_sectiontitle: { type: string }
        post_sectiondescription: { type: string }
        seo_settings:
          type: object
          properties:
            seo_image: { $ref: '#/components/schemas/ImageRef' }
            seo_title: { type: string }
            seo_description: { type: string }
            seo_keywords: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ProjectContent:
      type: object
      required: [id, title, company_name, project_description, project_image, project_overview, project_name, project_link, skills, ogAssetKey, ogImagePath, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        company_name: { type: string }
        project_description: { type: string }
        project_image: { $ref: '#/components/schemas/ImageRef' }
        project_overview:
          type: array
          items: { type: string }
        project_name: { type: string }
        project_link: { type: string }
        skills:
          type: array
          items: { $ref: '#/components/schemas/SkillEntry' }
        ogAssetKey: { type: string }
        ogImagePath: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    PostContent:
      type: object
      required: [id, title, author, post_text, post_image, publish_date, permalink, ogAssetKey, ogImagePath, createdAt, updatedAt]
      properties:
        id: { type: string }
        title: { type: string }
        author: { type: string }
        post_text: { type: string }
        post_image: { $ref: '#/components/schemas/ImageRef' }
        publish_date: { type: string }
        permalink: { type: string }
        ogAssetKey: { type: string }
        ogImagePath: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    HomeResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/HomeContent' }
    ProjectsResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: array
          items: { $ref: '#/components/schemas/ProjectContent' }
    ProjectResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/ProjectContent' }
    PostsResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data:
          type: array
          items: { $ref: '#/components/schemas/PostContent' }
    PostResponse:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: { $ref: '#/components/schemas/PostContent' }
    HomeUpdateInput:
      type: object
      additionalProperties: true
      description: Partial home document patch object.
    ProjectCreateInput:
      type: object
      required: [title, project_description]
      properties:
        id: { type: string }
        title: { type: string }
        company_name: { type: string }
        project_description: { type: string }
        project_image: { $ref: '#/components/schemas/ImageRef' }
        project_overview:
          type: array
          items: { type: string }
        project_name: { type: string }
        project_link: { type: string }
        skills:
          type: array
          items:
            oneOf:
              - type: string
              - type: object
                properties:
                  id: { type: string }
                  skill_name: { type: string }
    ProjectPatchInput:
      type: object
      additionalProperties: true
      description: Non-empty project patch payload.
    PostCreateInput:
      type: object
      required: [title, post_text]
      properties:
        id: { type: string }
        title: { type: string }
        author: { type: string }
        post_text: { type: string }
        post_image: { $ref: '#/components/schemas/ImageRef' }
        publish_date: { type: string }
        permalink: { type: string }
    PostPatchInput:
      type: object
      additionalProperties: true
      description: Non-empty post patch payload.
...